extracted hardcoded stings from the page
