+++ frontend/next.config.mjs
@@ -7,6 +7,8 @@ const withBundleAnalyzer = bundleAnalyzer({
   enabled: process.env.ANALYZE === "true",
 });
 
+const withNextIntl = createNextIntlPlugin("./i18n.ts");
+
 const backendUrl = process.env.NEXT_PUBLIC_BACKEND_URL ?? 'http://localhost:4000'
 
 /**
@@ -35,4 +37,4 @@ const nextConfig = {
   ...performanceConfig,
 };
 
-export default withBundleAnalyzer(nextConfig);
+export default withBundleAnalyzer(withNextIntl(nextConfig));
